+++ src/commands/interaction/runtime/resolution.test.ts
@@ -10,7 +10,7 @@ import {
 import { resolveRecordedTarget } from '@agent-device/selectors';
 import { makeSnapshotState } from '../../../__tests__/test-utils/snapshot-builders.ts';
 import type { Point } from '@agent-device/kernel/snapshot';
-import { INTERACTION_ERROR_REASONS } from '@agent-device/contracts/interaction-error';
+import { INTERACTION_ERROR_REASONS } from '../../../core/interaction-error.ts';
 import {
   clickRefE2,
   coveredByTabBarSnapshot,
