+++ src/core/interaction-error.ts
