+++ src/daemon/adapters/maestro/__tests__/daemon-runtime-port-selector-fallback.test.ts
@@ -1,4 +1,4 @@
-import { INTERACTION_ERROR_REASONS } from '@agent-device/contracts/interaction-error';
+import { INTERACTION_ERROR_REASONS } from '../../../../core/interaction-error.ts';
 import type { DaemonError } from '@agent-device/kernel/errors';
 import { expect, test } from 'vitest';
 import type { DaemonRequest } from '../../../daemon-request.ts';
