+++ src/daemon/handlers/session-replay-command.ts
@@ -4,7 +4,7 @@ import type { SessionState } from '../session-state.ts';
 import type { SessionStore } from '../session-store.ts';
 import type { LeaseRegistry } from '../lease-registry.ts';
 import type { BindDeviceRuntime, InspectDeviceRuntimeFacts } from '../request-runtime-binding.ts';
-import type { PlatformResourceCleanup } from '@agent-device/contracts/platform-resource-cleanup';
+import type { PlatformResourceCleanup } from '../platform-resource-cleanup.ts';
 import { runReplayCommand, runReplayTestCommand, type ReplaySession } from '../replay/index.ts';
 import { handleSessionCloseCommands } from '../session-lifecycle/index.ts';
 import { createReplayTestVideoOwner } from './session-replay-video-owner.ts';
