+++ src/daemon/handlers/snapshot.ts
@@ -8,7 +8,7 @@ import { dispatchSnapshotViaRuntime } from '../snapshot-runtime.ts';
 import { dispatchWaitViaRuntime } from '../wait-runtime.ts';
 import { resolveSessionDevice, withSessionlessRunnerCleanup } from '../snapshot-session.ts';
 import type { BindDeviceRuntime, InspectDeviceRuntimeFacts } from '../request-runtime-binding.ts';
-import type { PlatformResourceCleanup } from '@agent-device/contracts/platform-resource-cleanup';
+import type { PlatformResourceCleanup } from '../platform-resource-cleanup.ts';
 
 type SnapshotCommandParams = {
   req: DaemonRequest;
