+++ src/daemon/platform-resource-cleanup.ts
