+++ src/daemon/request-handler-chain.ts
@@ -23,7 +23,7 @@ import type { ScreenRecordingAdmissionLedger } from './screen-recording-admissio
 import type { PlatformRequestScope } from '@agent-device/contracts/platform-runtime-host';
 import type { RequestPlatformProviderScope } from '@agent-device/contracts/platform-providers';
 import type { AndroidObservationAdapter } from '@agent-device/contracts/android-observation';
-import type { PlatformResourceCleanup } from '@agent-device/contracts/platform-resource-cleanup';
+import type { PlatformResourceCleanup } from './platform-resource-cleanup.ts';
 
 type RequestHandlerChainParams = {
   req: DaemonRequest;
