+++ src/daemon/session-lifecycle/internal/session-close-lifecycle-teardown.ts
@@ -13,7 +13,7 @@ import {
   stopSessionPerfCapture,
   type SessionCleanupFailure,
 } from '../../session-teardown.ts';
-import type { PlatformResourceCleanup } from '@agent-device/contracts/platform-resource-cleanup';
+import type { PlatformResourceCleanup } from '../../platform-resource-cleanup.ts';
 import { hasRuntimeTransportHints, runtimeHintValues } from '../../session-runtime.ts';
 import type {
   CloseRuntime,
