+++ src/daemon/snapshot-runtime-binding.ts
@@ -17,7 +17,7 @@ import {
   type InspectDeviceRuntimeFacts,
 } from './request-runtime-binding.ts';
 import type { DeviceReadyOptions } from './device-ready.ts';
-import type { PlatformResourceCleanup } from '@agent-device/contracts/platform-resource-cleanup';
+import type { PlatformResourceCleanup } from './platform-resource-cleanup.ts';
 import { SessionStore } from './session-store.ts';
 import type { DaemonRequest, DaemonResponse } from './daemon-request.ts';
 import type { SessionState } from './session-state.ts';
